requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
email-validator>=2.2.0
Flask-Migrate>=4.0.7
Werkzeug>=3.0.0
# Werkzeug>=2.3.7
Pillow>=10.4.0
python-dotenv>=1.0.1
WTForms==3.0.1